backend/requirements.txt
aniso8601==8.0.0
cffi==1.14.0
click==7.1.1
cryptography==2.8
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.7
Flask-SocketIO==4.2.1
Flask-Sockets==0.2.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.10.3
python-engineio==3.13.0
pytz==2019.3
simplejson==3.17.0
six==1.14.0
Werkzeug==1.0.1
sendgrid==6.4.7
Twisted==20.3.0
autobahn==20.6.1
Flask-Cors==3.0.8
Markdown==3.2.2